pyproject.toml
[project]
name = "reloaper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiopubsub>=3.0.0",
    "attrs>=24.2.0",
    "numpy>=2.1.3",
    "prompt-toolkit>=3.0.48",
    "pynput>=1.7.7",
    "sounddevice>=0.5.1",
    "sunvox-dll-python",
    "typer>=0.13.1",
    "watchfiles>=1.0.0",
]

[tool.uv.sources]
sunvox-dll-python = { git = "https://github.com/metrasynth/sunvox-dll-python", branch = "master" }
